Game Loop Start Game Timer working
Use of new C++ Player class allowed movement. Ensured that the system currently works for 1 player HUD.
Next steps.  On completion of the Timer, run a GAME_HAS_STARTED function.

After the game has started, we need a lose condition and a way to return to the beginning state.
Use of Levels?