Boundary Box Size and Floor Height Check
The size of the boundary box X and Y Coords are shown in the level editor. So you know the spawn area.
The Spawned enemy Class is defined in the Spawner blueprint.
The Height is also offset by a Vec3. This is broken out with comments in the BP_EnemySpawner.